requirements.txt
Django==2.1.8
django-filter==2.0.0
djangorestframework==3.9.0
pycparser==2.19
pycryptodome==3.6.6
PyMySQL==0.9.2
redis==2.10.6
requests==2.20.1
urlparser==0.1.2
